Remove redundant implementations of delay
